# docs/annotation_openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Construction Safety Annotation API",
    "description": "Free-form GroundingDINO detection, SAM box-prompt segmentation, Qwen prompt enrichment, manual annotation review, and dataset release API. The API process does not load model weights.",
    "version": "1.5.1"
  },
  "paths": {
...
    "/v1/annotation/task-batches/mask-candidates": {
      "post": {
        "tags": [
          "Task Batches"
        ],
        "summary": "Create Batch Mask Candidates",
        "operationId": "createBatchMaskCandidates",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchMaskCandidatesRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "One SAM operation is accepted per valid Task. Rejected items do not prevent other Tasks from being queued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchOperationsAccepted"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "422": {
            "description": "Batch request structure is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "503": {
            "description": "Storage is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/v1/annotation/task-batches/prompt-enrichments": {
      "post": {
        "tags": [
          "Task Batches"
        ],
        "summary": "Create Batch Prompt Enrichments",
        "operationId": "createBatchPromptEnrichments",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPromptEnrichmentsRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "One Qwen operation is accepted per valid Task. Rejected items do not prevent other Tasks from being queued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchOperationsAccepted"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "422": {
            "description": "Batch request structure is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "503": {
            "description": "Storage is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/annotation/tasks/{task_id}/mask-candidates": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create Mask Candidate",
        "operationId": "createMaskCandidate",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaskCandidateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "SAM mask generation operation accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAccepted"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "404": {
            "description": "Task was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "503": {
            "description": "Storage is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict or invalid state transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "422": {
            "description": "Annotation payload is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
    "/v1/annotation/tasks/{task_id}/prompt-enrichments": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create Prompt Enrichment",
        "operationId": "createPromptEnrichment",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePromptEnrichmentRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Qwen2.5-VL prompt generation operation accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationAccepted"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "404": {
            "description": "Task was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "503": {
            "description": "Storage is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "409": {
            "description": "Version conflict or invalid state transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          },
          "422": {
            "description": "Annotation payload is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorPayload"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchMaskCandidateItem": {
        "properties": {
          "task_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Task Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "box_xyxy": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Box Xyxy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "expected_version",
          "box_xyxy"
        ],
        "title": "BatchMaskCandidateItem"
      },
      "BatchMaskCandidatesRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchMaskCandidateItem"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "BatchMaskCandidatesRequest"
      },
      "BatchOperationItemResult": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "operation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "rejected"
            ],
            "title": "Status"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorPayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "BatchOperationItemResult"
      },
      "BatchOperationsAccepted": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchOperationItemResult"
            },
            "type": "array",
            "title": "Items"
          },
          "accepted_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Accepted Count"
          },
          "rejected_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Rejected Count"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accepted_count",
          "rejected_count"
        ],
        "title": "BatchOperationsAccepted"
      },
      "BatchPromptEnrichmentItem": {
        "properties": {
          "task_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Task Id"
          },
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "expected_version"
        ],
        "title": "BatchPromptEnrichmentItem"
      },
      "BatchPromptEnrichmentsRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchPromptEnrichmentItem"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "BatchPromptEnrichmentsRequest"
      },
...
      "CreateMaskCandidateRequest": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "box_xyxy": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Box Xyxy"
          },
          "boxes_xyxy": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "maxItems": 128,
            "minItems": 1,
            "title": "Boxes Xyxy"
          },
          "detection_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Detection Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "CreateMaskCandidateRequest"
      },
      "CreatePromptEnrichmentRequest": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "CreatePromptEnrichmentRequest"
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "validation_error",
          "annotation_validation_failed",
          "unauthorized",
          "forbidden",
          "not_found",
          "version_conflict",
          "idempotency_conflict",
          "invalid_state_transition",
          "unsupported_media_type",
          "request_too_large",
          "queue_full",
          "model_unavailable",
          "downstream_unavailable",
          "internal_error"
        ],
        "title": "ErrorCode"
      },
      "ErrorDetail": {
        "properties": {
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "ErrorDetail"
      },
      "ErrorPayload": {
        "properties": {
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorPayload"
      },
...
      "OperationAccepted": {
        "properties": {
          "operation_id": {
            "type": "string",
            "title": "Operation Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running"
            ],
            "title": "Status",
            "default": "queued"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "operation_id",
          "created_at"
        ],
        "title": "OperationAccepted"
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}